Extract the fake ConfigurationPanelForm class used in the unit test
